Department addition

+++ src/main/java/org/gklsan/springboot/myjavalearning/repository/StudentRepository.java
@@ -2,9 +2,8 @@ package org.gklsan.springboot.myjavalearning.repository;
 
 import org.gklsan.springboot.myjavalearning.entity.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 // custom routes names
 //@RepositoryRestResource(path = "people")
 
-public interface StudentRepository extends JpaRepository<Student, Integer> {}
+public interface StudentRepository extends JpaRepository<Student, Long> {}
